Disabling job folder deletion on job deletion

+++ brender/server/application/modules/jobs/__init__.py
@@ -378,9 +378,12 @@ class JobApi(Resource):
             TaskApi.delete_tasks(job.id)
             TaskApi.create_tasks(job)
 
+            #Security check
+            insecure_names=[None, "", "/", "\\", ".", ".."]
             path = os.path.join(job.project.render_path_server, str(job.id))
-            if os.path.exists(path):
-                rmtree(path)
+            if job.project.render_path_server not in insecure_names and str(job.id) not in insecure_names:
+                if os.path.exists(path):
+                    rmtree(path)
             logging.info('Job {0} reset end ready'.format(job_id))
 
 class JobDeleteApi(Resource):
@@ -395,10 +398,10 @@ class JobDeleteApi(Resource):
             if job:
                 path = os.path.join(job.project.render_path_server, str(j))
                 #Security check
-                insecure_names=[None, "", "/", "\\", ".", ".."]
-                if job.project.render_path_server not in insecure_names and str(j) not in insecure_names:
-                    if exists(path):
-                        rmtree(path)
+                #insecure_names=[None, "", "/", "\\", ".", ".."]
+                #if job.project.render_path_server not in insecure_names and str(j) not in insecure_names:
+                #    if exists(path):
+                #        rmtree(path)
 
                 db.session.query(JobManagers).filter(JobManagers.job_id == job.id).delete()
                 db.session.delete(job)
